Create Applet
Develop an applet that draws a circle. The dimension of the applet should be 500 x 300 pixels. The circle should be centered in the applet and have a radius of 100 pixels. Display your name centered in a circle (using drawOval() method).

Firstly, the dimensions of the applet should be 500 x 300, so I have to define attributes height = 300 and width = 500 in the applet code tag.
Second of all, I have used the drawOvel() method here, It is a method of Graphics class, it has 4 parameters x, y, width, and height. We have to draw a circle in the center, the applet size is 500 x 300 so the center coordinates will be 250 and 150. And radius is 100 so starting point where we can start drawing the circle is 150 and 50 (250-100 will be x coordinate and 150-100 will be y), the radius is 100 so height and width will be double of it which means height will be equal to the width in order to draw a circle and that is 200.
Lastly, for writing the name I have used the drawString() method which is a method of Graphics class, it has 3 parameters, String to be written, x, and y. Here name should be in the center of the circle, so x and y will be 250 and 150 (applet size is 500 x 300 so the center coordinates will be 250 and 150).